+++ behave/formatter/json.py
@@ -32,7 +32,7 @@ class JSONFormatter(Formatter):
         self.current_feature_data = None
         self._step_index = 0
 
-
+    # -- FORMATTER API:
     def uri(self, uri):
         pass